refactor(.): setup routes and architecture
